requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
werkzeug==3.0.4